// StockRadarFrontend/src/company.d.ts
export interface CompanySearch {
  currency: string;
  exchangeShortName: string;
  name: string;
  stockExchange: string;
  symbol: string;
}